Restructure pkg/ folder in preparation of new parser
