+++ pkg/tf-profile/core/errors.go
@@ -1,0 +1,9 @@
+package tfprofile
+
+type (
+	LineParseError struct{ Msg string }
+)
+
+func (e *LineParseError) Error() string {
+	return e.Msg
+}
