+++ pkg/tf-profile/readers/stdin_reader.go
