+++ pkg/tf-profile/stats/stats.go
@@ -5,7 +5,11 @@ import (
 	"fmt"
 	"sort"
 
+	. "github.com/QuintenBruynseraede/tf-profile/pkg/tf-profile/aggregate"
 	. "github.com/QuintenBruynseraede/tf-profile/pkg/tf-profile/core"
+	. "github.com/QuintenBruynseraede/tf-profile/pkg/tf-profile/parser"
+	. "github.com/QuintenBruynseraede/tf-profile/pkg/tf-profile/readers"
+	. "github.com/QuintenBruynseraede/tf-profile/pkg/tf-profile/utils"
 	"github.com/fatih/color"
 	"github.com/rodaine/table"
 )
